+++ src/router.go
@@ -1,0 +1,19 @@
+package main
+
+import "github.com/gorilla/mux"
+
+func newRouter() *mux.Router {
+	router := mux.NewRouter().StrictSlash(true)
+	for _, route := range routes {
+		router.
+			Methods(route.Method).
+			Path(route.Pattern).
+			Name(route.Name).
+			Handler(route.HandlerFunc)
+	}
+	return router
+}
+
+/*
+ * END OF FILE!
+ */
